refactor(uikit): update styles and create hooks
