+++ src/components/BarChart/BarChartTypes.ts
@@ -15,3 +15,15 @@ export interface IBarChart extends IChart {
   currentFilter?: string;
   handleChangeFilter?(filter: string): void;
 }
+
+export type BarChartHookData = {
+  chartData: any;
+  keyNames: string[];
+  missingData: boolean;
+};
+
+export type BarChartHookCallbacks = {
+  isHide: (keyName: string) => boolean;
+  setHide: React.Dispatch<React.SetStateAction<Record<string, boolean>>>;
+  getCoordinates: (max: number, gap: number) => number[];
+};
