+++ src/components/BarChart/index.tsx
@@ -84,21 +84,8 @@ const Chart = ({
     <>
       <div className="bar-chart-graphic-container">
         {dateFilter && (
-          <div
-            className="display-flex"
-            style={{
-              width: "100%",
-              justifyContent: "end",
-            }}
-          >
-            <div
-              className="display-flex"
-              style={{
-                alignItems: "end",
-                width: "40%",
-                justifyContent: "center",
-              }}
-            >
+          <div className={clsx("display-flex", "filter-container")}>
+            <div className={clsx("display-flex", "filter-card")}>
               <Tabs
                 value={currentFilter}
                 centered
@@ -215,7 +202,7 @@ const Chart = ({
         </ResponsiveContainer>
       </div>
       <div className={clsx("display-flex", "hide-bars-container")}>
-        <div className="display-flex" style={{ gap: "16px" }}>
+        <div className={clsx("display-flex", "checkbox-container")}>
           {(keyNames || []).map((name, index) => (
             <span
               className={clsx("flex-center", "variant-body1")}
