+++ src/scss/components/DropDown/_dropdown.scss
@@ -11,6 +11,10 @@
   border: 1px solid #26a66b;
 }
 
+.help-text {
+  color: #7c7d7e;
+}
+
 .checkBox {
   .css-12wnr2w-MuiButtonBase-root-MuiCheckbox-root.Mui-checked {
     color: #26a66b;
